The engine informs now with which container a character is interacting.
The current interaction target of a character (e.g. a container) gets now
stored within a new CharacterComponent field called interactionTarget.

A new utility class CharacterUtil can be used to easly specify the current
interaction target of a character.

At the start of an interaction an InteractionStartEvent gets sent to
the target entity (e.g. the container).
Systems can listen to this event to open certain UI dialogs.
As the event is sent to the target, it is possible to filter for
interaction events of entities with certain components.
This event replaces the old OpenInventoryEvent.

At the end of an interaction an InteractionEndEvent gets sent.
Systems can listen to this event to close certain UI dialogs.